Move uinput stuff to UinputDevice
